requirements.txt
selenium==4.15.1
pytest>=6.0.0, <7.0.0
pytest-selenium==4.0.0
webdriver-manager==3.7.0
allure-pytest